Filters in R
That is exactly what the filter function does. It selects or filters the rows of the data table that meet certain criteria creating a new data subset. Filter is a function of the dplyr package so, in order to use it, you need to have installed and loaded the dplyr package.