Association rules Titanic in r
The Titanic data set provides information on the fate of passengers on the fatal maiden voyage of the ocean liner ‘Titanic’, summarized according to economic status (class), sex, age, and survival. These data were originally collected by the British Board of Trade in their investigation of the sinking. Titanic, which is available in the R package datasets