PCA Algorithm
In statistics, principal component analysis is a method to project data in a higher-dimensional space into a lower-dimensional space by maximizing the variance of each dimension. Given a collection of points in two, three, or higher-dimensional space, a "best fitting" line can be defined as one that minimizes the average squared distance from a point to the line.