Principal Component Analysis  using Wine dataset
Principal Components Analysis (PCA) is an algorithm to transform the columns of a dataset into a new set of features. A large chunk of the information across a large dataset can effectively be compressed in fewer columns. This enables dimensionality reduction and the ability to visualize the separation of classes or clusters if any.