Simple Linear Regression With Salary Dataset
Simple Linear Regression, which is a very basic of Linear Regression algorithm frequently used in Machine Learning, is applied in solving real-life problems. For this purpose, I'll use a dataset in which I have to predict the Salary of an individual(say, a person) on the basis of one explanatory variable Years of Experience.