Multiple Time Series
time-series is a series of data points in which each data point is associated with a timestamp. A simple example is a price of a stock in the stock market at different points of time on a given day. Another example is the amount of rainfall in a region at different months of the year. R language uses many functions to create, manipulate, and plot the time series data. The data for the time series is stored in an R object called a time-series object. It is also an R data object like a vector or data frame.